Changed CSS links to original calls
It seems when I zip the folder all the CSS filenames revert to their original name which started with a capital letter, so by changing the calls it should resolve the issue. I'm not sure whether this has something to do with the compression algorithm of Zip but it took me a while to spot.